backend/pyproject.toml
[tool.poetry]
name = "manz"
version = "0.1.0"
description = "A simple meal planner"
authors = ["Antony Gandonou migan <gmigan.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
black = "^24.10.0"
ruff = "^0.8.6"
bandit = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
